Change reference to CXXFLAG to CXXFLAGS
This typo is the source of a variety of issues I have had upgrading packages to
use this version of common.mk. Suprisingly atleast 4 projects were able to build
correctly with this bug, so it might not be everywhere. I will check all of the
packages I have attempted to update for this.

BUG=chromium-os:37854
TEST=Copy common.mk into a variety of packages that depend on it and confirm
     that they build correctly with USE=profiling set

Change-Id: I2e344db744c89b40974afa9cf4880f0b4c2ff112
Reviewed-on: https://gerrit.chromium.org/gerrit/42869
Tested-by: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Ryan Harrison <rharrison@chromium.org>