shill: Add netfilter queue processor shim
This user-space process is repsonsible for returning a verdict on
incoming UDP packets that would otherwise be dropped.  A netfilter
queue is created using iptables to select incoming frames for which
this verdict is requested.  This processor determines the verdict
based on recent outgong multicast packets that are selected based
on a separate iptables queue.

Using this technique, we can track output frames and accept replies
to multicast request that arrives within a short time window after
it is sent.

CQ-DEPEND=CL:43179, CL:43133
BUG=chromium-os:38605
TEST=Unit tests + manually tested using a hand-built multicast
requester/responder pair.

Change-Id: Ib5ec2a22aa020b8a0588ac7a6bcfa31743b3ee7a
Reviewed-on: https://gerrit.chromium.org/gerrit/43096
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>