Crosh: Use the new try_touch_experiment script
Crosh used to have it's own shell script version of try_touch_experiment,
but there is now a program that provides a much nicer user experience.
This removes the old version of the command in Crosh and replaces it with
a call out to the newly installed one instead.

BUG=chromium-os:38637
TEST=Manually tested on Link
CQ-DEPEND=Ia0090629e8d988e8f517ff83ceac03a9bf808c19

Change-Id: I1bc84453f25e088d3ac5960d54f2c2eb139617f4
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/43185
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>