power: Put backlight into suspended state earlier.
This ensures that powerd puts the backlight into the
"suspended" state (i.e. original level saved to resume level
and current level set to 0) before announcing to Chrome that
the system is going to suspend (which causes Chrome to turn
the display on so it will be turned on immediately upon
resuming).

BUG=chromium-os:38869
TEST=updated unit tests

Change-Id: I96b310f6c03fed1b3e42c9578605ce14d3214d45
Reviewed-on: https://gerrit.chromium.org/gerrit/43257
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Queue: Daniel Erat <derat@chromium.org>
Tested-by: Daniel Erat <derat@chromium.org>