shill: Add DBus interfaces to verify destinations
Add the DBus API Chrome will use to call to shill to verify destinations
and encrypt WiFi credentials.  This includes 3 new manager interface
methods: one to verify credentials, one to verify and encrypt a
passphrase for the specified service, and one to verify and encrypt a
specified string of data.

BUG=chromium-os:38993
TEST=Compiles, these are just stubs for now.

Change-Id: I63aa1f9faf272444dd8b844c5a8fdc000ee813a5
Reviewed-on: https://gerrit.chromium.org/gerrit/43457
Tested-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Christopher Wiley <wiley@chromium.org>