Fix a clang warning about sprintf behavior we need in this package.
inst_util.cc accepts a format string and then passes it to vsnprintf. clang
warns about this, even though gcc failed to do so. Add an attribute to
the helper method so that the compiler knows it's a printf style method
and can do format checking.

I have tested that this still builds with gcc, but I don't know if it actually
fixes the warning with clang.

BUG=chromium-os:35541
TEST=Built (since the change is to a build warning, this seems sufficient.)

Change-Id: I965a77dbfa2090b109fb43eddbb79b7cc858750c
Reviewed-on: https://gerrit.chromium.org/gerrit/43588
Reviewed-by: Guozhi Wei <carrot@google.com>
Reviewed-by: Gilad Arnold <garnold@chromium.org>
Commit-Queue: Don Garrett <dgarrett@chromium.org>
Tested-by: Don Garrett <dgarrett@chromium.org>