power: Increase logging during AmbientLightSensorTest.
This makes AsyncFileReader log more details during
AmbientLightSensorTest, with the hope that it'll make it
easier to track down the cause of "corrupted double-linked
list" crashes in glibc that occasionally show up on the
builders.

I'm also fixing a (probably harmless)
use-before-initialization in AmbientLightSensor.

BUG=chromium-os:38609
TEST=ran tests

Change-Id: Ie860b0137e52c71da31138ca50c7fe9a675c5288
Reviewed-on: https://gerrit.chromium.org/gerrit/43613
Commit-Queue: Daniel Erat <derat@chromium.org>
Reviewed-by: Daniel Erat <derat@chromium.org>
Tested-by: Daniel Erat <derat@chromium.org>
Reviewed-by: Ryan Harrison <rharrison@chromium.org>