Add DarkPolicy class and unittests
This class contains all of the functionality to keep track of preferences for
and determine how long the system suspends for. It also determines when we
should shutdown based on battery state. With the exception of the constructor,
the interface is designed so the caller does not know what other classes may be
used to determine these things.

BUG=chromium-os:39284
TEST=Suspend system

Change-Id: Ifaa4187d838bb89c0525900534761e3120bc9778
Reviewed-on: https://gerrit.chromium.org/gerrit/43874
Reviewed-by: Derek Basehore <dbasehore@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>