Use dark resume and move parts of powerd_suspend
This makes the suspender class use dark resume to resuspend or shutdown if we
perform a dark resume. It also moves muting the system and sending out the
PowerStateChanged signal through the dbus to powerd. This allows the suspender
class to call powerd_suspend again after waking up from a dark resume without
messing with other programs through the PowerStateChanged signal. It also lets
us not store state in some file for whether the system was muted for when we
call powerd_suspend again.

This also modifies suspender_unittest.cc to take into account the changes to
suspender.cc

BUG=chromium-os:39284
TEST=Suspend system

Change-Id: I4d5ff09c277c22c2324f59e8e9b7bb0cc7c54ce7
Reviewed-on: https://gerrit.chromium.org/gerrit/43876
Reviewed-by: Derek Basehore <dbasehore@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>