power: Update suspend_stress_test to ignore a specific firmware error.
suspend_stress_test greps the firmware log for the string "ERROR".
Unfortunately, we have a print in Coreboot that contains the string
"ERROR", yet is not a fatal error or an error at all. It causes
suspend_stress_test to prematurely abort on certain platforms.

Fix this by ignoring this one specific error string. We are also
changing the string in Coreboot for future firmware builds.

TEST=Manual. Verify suspend_stress_test doesn't abort on Stout.
BUG=chrome-os-partner:17973.

Change-Id: I982f965b9e8513071276008c7e0f69f07d381360
Reviewed-on: https://gerrit.chromium.org/gerrit/43920
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-by: Daniel Erat <derat@chromium.org>
Commit-Queue: Shawn Nematbakhsh <shawnn@chromium.org>
Tested-by: Shawn Nematbakhsh <shawnn@chromium.org>