common-mk: Added in test to profiling rule for empty file list
For projects that do not have tests that are run with `make test` the profiling
phase of make test will fail since gcov doesn't know what to do with an empty
file list. This causes an issue, because for the cros_workon.eclass there is a
test rule that may be run, so it makes sense to make common.mk more robust.

BUG=chromium-os:39210
TEST=Built a package without implemented tests with tests turned on to confirm
     that the profiling step passed without doing anything.
     Built a package with implemented tests with tests turned on to confirmed
     profiling occured as expected.

Change-Id: Ia51179a721198c4615911b489fbd50f7289abcb5
Reviewed-on: https://gerrit.chromium.org/gerrit/43928
Commit-Queue: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Ryan Harrison <rharrison@chromium.org>
Tested-by: Ryan Harrison <rharrison@chromium.org>