init: update preload-network script for ath9k_btcoex
Previously, the logic for finding the module for a network
device assumed that the driver output by ethtool had the
same name as the module that provided it.

While this is normally the case, it is not true for ath9k_btcoex.
In that case, the driver name is "ath9k", but the module name
is "ath9k_btcoex".

Update the logic to use the /sys entry for the network device,
which provides the name of the module.

BUG=chromium-os:38958
TEST=manual (see below)

Manual testing
--------------
 1. use a link, in dev mode
 2. rm /var/lib/preload-network-drivers
 3. reboot
 4. wait for login prompt
 5. connect to wifi
 6. cat /var/lib/preload-network-drivers
    -> this should say "ath9k_btcoex"
 7. reboot
 8. wait for login prompt
 9. connect to wifi
10. grep "phy0: ASPM" /var/log/messages | tail -2
    -> observe that the last line has a kernel timestamp
       approximately 2 seconds early than the first line.
       e.g.
       <syslog header> [    6.729392] ath: phy0: ASPM ...
       <syslog header> [    4.723810] ath: phy0: ASPM ...

Change-Id: Ieeb86475ec24b12122cd1dbe7dcb3ec7e8d34614
Reviewed-on: https://gerrit.chromium.org/gerrit/43989
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>