Create a directory in which Chrome can store a reboot request
When an automatic reboot is scheduled, Chrome needs to persist
the reboot request in a file so that if the Chrome process
crashes, the request is not lost. The file needs to reside in
a tmpfs so that once the reboot does occur, the file is
guaranteed to be destroyed and no reboot loop can occur.

BUG=chromium-os:36400
TEST=Manual; directory is created and has the right permissions

Change-Id: I9a251caaa5562b61d4564ed9f69385a65dea16c3
Reviewed-on: https://gerrit.chromium.org/gerrit/44077
Tested-by: Bartosz Fabianowski <bartfab@chromium.org>
Reviewed-by: Daniel Erat <derat@chromium.org>
Commit-Queue: Bartosz Fabianowski <bartfab@chromium.org>