Add 'sound' crosh command to enable/disable beam forming on Pixel.
Use:
> sound beamforming <on|off>

This will enable or disable the beam forming for the built-in mic
array on Pixel.  It will be used to evaluate the feature while we
decide if a UI element is needed to toggle it.

BUG=chromium-os:39215
TEST=run command and check that value is set with amixer on Link.
Check that an error is returned on Lumpy where the codec doesn't
support it.

Change-Id: I4dba7a1303541c6f1ba3b75783f63b3fe26eb03f
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/44147
Reviewed-by: Chih-Chung Chang <chihchung@chromium.org>
Reviewed-by: Kees Cook <keescook@chromium.org>