Implemented a PKCS #11 backed key store.
Certified keys which are associated with the user need a safe place to
live.  The most important requirement is that the key must not be
available when the user is not signed in.  This CL implements a key
store using data objects in the user's PKCS #11 token.

BUG=chromium-os:38996
TEST=unit
CQ-DEPEND=CL:44332, CL:44334

Change-Id: Ice506b0aed92137eef99150ac6f7d5ecf04ce1af
Reviewed-on: https://gerrit.chromium.org/gerrit/44336
Commit-Queue: Darren Krahn <dkrahn@chromium.org>
Reviewed-by: Darren Krahn <dkrahn@chromium.org>
Tested-by: Darren Krahn <dkrahn@chromium.org>