power_manager: Reset hysteresis state when returning to active state.
Since ALS based transitions are supressed while we are in a non-active state
like idle some of the transitions maybe missed. This leads to cases where if the
ambient light levels change while we are non-active the backlight might not do
what the user expects when becoming active. The solution for this is reset the
state of the hysetersis, so that we can run what ever transition is approriate
now.

BUG=chrome-os-partner:17935
TEST=On link device let device idle in bright room, so the backlight is
     off. Cover the ALS sensor, then interact witht eh device so it comes out of
     idle. The expected behaviour is the backlight should turn on again.

Change-Id: Ib3f45442b016ffec283226be78777d742b4716c4
Reviewed-on: https://gerrit.chromium.org/gerrit/44206
Reviewed-by: Daniel Erat <derat@chromium.org>
Tested-by: Ryan Harrison <rharrison@chromium.org>
Commit-Queue: Ryan Harrison <rharrison@chromium.org>