shill: Manager: Create Manager::ConfigureServiceForProfile
Create a method in the manager for creating a WiFi service
within a different profile than the one the service currently
exists without affecting the current profile entry.

To accomplish this, add methods in the WiFi Provider to
return a matching service given a KeyValueStore, as well
as the ability to create a temporary service for configuration
purposes only.  While here, allow ConfigureService to accept
AdHoc mode WiFi services.

BUG=chromium-os:39513
TEST=Unit tests

Change-Id: I4d0fc9b41cf6a5d7da59a87ef58ee5d491bd43a6
Reviewed-on: https://gerrit.chromium.org/gerrit/44563
Commit-Queue: Paul Stewart <pstew@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>