power_manager: Reset all of ALS state when becoming active & initialized
This changes the name of the reset method from ResetHysteresis ->
ResetAls. Additionally this function now sest the ALS level to the current level
of the device and adjusts related variables.

This is needed, because when we become active the system is going to apply the
current ALS level. Currently the code just reapplies the old value, but if light
was on before, but should no longer be on, then the light will go
dim->on->off. By setting the ALS level to the dim level we are effectively
transitioning dim->dim->off in this case. There is also an issue with the
backlight being sticky in this situation that this fix resolves.

BUG=chrome-os-partner:17935
TEST=On a link device let device idle in a dark enough room for the backlight to
     be on. Take into a bright environment, then interact with the device so it
     comes out of idle. The expected behaviour is the backlight should make a
     single transition from the dim state to off if it isn't already off. Repeat
     starting bright and going to dark. There should be a single transition from
     dim/off to on.

Change-Id: I84541b12834f4381ec7ed228049c3cea04cb2c18
Reviewed-on: https://gerrit.chromium.org/gerrit/44655
Commit-Queue: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Ryan Harrison <rharrison@chromium.org>
Tested-by: Ryan Harrison <rharrison@chromium.org>