shill: Removes nl80211_socket but NOT YET handles broadcast messages.
This CL removes the use of libnl for sending and receiving netlink
messages.  This required replacing all of the NetlinkSocket code and
enabled the removal of Nl80211Socket code.  There is some procesing
required to get broadcast messages that has not been implemented in this
CL (that's the subject of the next CL) so not all functionality is
provided, here.  In addition to that and a few obvious changes, this CL
does the following:

  1. Removes nl80211_socket -- it's no longer needed.
  2. Removes |AddGroupMembership| since that used to be a function of
     nl80211_socket; the functionality will be replaced in the next CL.
  3. Replaces mock_nl80211_socket.h with mock_netlink_socket.h (and
     substituted the a few method names).

Note: It's best to view the changes in netlink_socket.* as a total
replacement rather than a individual changes.  Look at it as entirely
new code.

CQ-DEPEND=CL:I04571bdff9908ad8bd39a3a64a24e7e11074bf18
BUG=chromium:218211
TEST=unittests

Change-Id: Ia24d6f02a34c92d62962e07b71c0cf54f98a32e4
Reviewed-on: https://gerrit.chromium.org/gerrit/44769
Commit-Queue: Wade Guthrie <wdg@chromium.org>
Reviewed-by: Wade Guthrie <wdg@chromium.org>
Tested-by: Wade Guthrie <wdg@chromium.org>