Make sure all perf samples have the same fields
perf samples have a packed-field format (see perf_event.h).
Whatever fields are not present in the sample don't have a block of
memory reserved for them.  The attrs.sample_type is a bitfield that
indicates which fields are present.

This patch adds a check to make sure that is always the case.  The CHECK
will trigger if that changes in a newer version of perf.

Also applies the check to protobufs in the serializer.

BUG=none
TEST=unit tests pass

Change-Id: If2199b38e895ecfb380a9d80346d01be286408b2
Signed-off-by: Simon Que <sque@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/44713
Reviewed-by: asharif <asharif@chromium.org>