power: Avoid suspending while the system is being updated.
This makes powerd listen for StatusUpdate D-Bus signals
emitted by the update engine so it can avoid suspending the
system while an update is being downloaded or applied (only
on AC power).

It also adds a helper method for some delay-triggered code
that was copied-and-pasted across a bunch of tests.

BUG=chromium-os:20502,chromium-os:39706,chromium-os:39700
TEST=added a unit test; also manually checked that idle
     action went to no-op while downloading and applying an
     update and returned to suspend after it was applied,
     and also returned to suspend eventually after pulling
     the network connection mid-download

Change-Id: I5b045ab7ccfef99f425b50d96438b44d0f1277c2
Reviewed-on: https://gerrit.chromium.org/gerrit/44873
Reviewed-by: Daniel Erat <derat@chromium.org>
Tested-by: Daniel Erat <derat@chromium.org>
Commit-Queue: Daniel Erat <derat@chromium.org>