Add bt_console command
This provides access to the Bluetooth console D-Bus client, not
direct access to the Bluetooth hardware.

Almost all features of the client are available within the Chromium
UI itself, except for:

  discoverable <on/off>   Set controller discoverable mode

Which is not present in the UI since we don't offer services at this
time, and did not want a local information leak without user awareness
(enabling discoverability via an advanced console command can be
definitely considered awareness).

The console also provides additional information for the adapter,
discovered and paired devices, that is not available in the UI. This
information consists of the Class of device, exported service UUIDs,
and pairing-related properties. These are often required for
debugging.

BUG=chromium-os:38832
TEST=run on device

Change-Id: I1b42a1f766f661b4e83cdbb760bd79a3c15b16b4
Reviewed-on: https://gerrit.chromium.org/gerrit/45085
Commit-Queue: Scott James Remnant <keybuk@chromium.org>
Reviewed-by: Scott James Remnant <keybuk@chromium.org>
Tested-by: Scott James Remnant <keybuk@chromium.org>