cryptohome: add support for multiple mounts to be used
Cryptohome supports user-specific mount points, but the DBus interface and
daemon do not have a good way to track multiple mounts. This change tweaks the
overall behavior to make Mounts act as per-user in-memory state only and have
cryptohomed manage the mount mappings explicitly.  Additionally, PKCS#11
initialization is restricted to the first mount and the mounttask is now
cancelable.  This is mostly just clean up work finishing what ellyjones@
started last year.

The resumption behavior should look similar to what cryptohomed does today with
the caveat that if a mount point has open file references when cryptohome
restarts and still has those when an over-mount request comes in,
Mount::MountCryptohome() will reject it as a busy mount point.
For this to occur, it comes to:
1. c-home crashes while signed in and for some reason the user tries to
   over-mount
2. c-home crashes during sign-out failing to unmount and ui.conf fails to
   remove references to the mount point
Both of these cases imply a behavior change or cascading failure.  Should we force unmount? reboot?

BUG=chromium-os:39682
TEST=(1)unittests pass;
Lumpy ToT test image:
(2) suite:smoke passes
(3) manual sign in as guest, call
    cryptohome --action=mount --create --ensure_ephemeral --user=a1 --password=a1
    cryptohome --action=mount --create --ensure_ephemeral --user=a2 --password=a2
    mount  # Shows our mounts
    initctl restart cryptohomed
    mount
    grep cryptohome /var/log/messages # Shows the top two mounts cleaned up and guestfs left alone
(4) manual sign in as a real user, call
    cryptohome --action=mount --create --user=a1 --password=a1
    cryptohome --action=mount --create --ensure_ephemeral --user=a2 --password=a2
    mount  # Shows our mounts
    initctl restart cryptohomed
    mount
    grep cryptohome /var/log/messages # Shows the top two mounts cleaned up and _all_ my mounts left alone
(5) manual sign-in as guest, call
      sudo bash
      # bash -c 'initctl stop cryptohomed; initctl restart ui; initctl start cryptohomed'
    sign in as guest:
      grep cryptohome /var/log/messages # see that stale guestfs mount was cleaned up
(6) #5, except sign-back in as a user
(7) #6, except sign in the whole time as the same real user

Change-Id: I022e99df2f2aea80d600ba85066ee93cdaf34027
Reviewed-on: https://gerrit.chromium.org/gerrit/44972
Reviewed-by: Elly Jones <ellyjones@chromium.org>
Commit-Queue: Will Drewry <wad@chromium.org>
Tested-by: Will Drewry <wad@chromium.org>