power: fix timeout leak in internal_backlight_unittest
In unittest InternalBacklightTest::Transitions we call the
transition timeout callback explicitly using
TriggerTransitionTimeoutForTesting(). In that case, the return
value from the callback HandleTransitionTimeout() is not actually
returned to the glib mainloop, so we may not cancel the timeout
while we finish the test. And the callback can be executed while
following tests are running which will cause very confusing failure
in the following tests. This might also contribute to
http://crosbug.com/38732

BUG=chromium-os:38732
TEST=unittests don't crash.

Change-Id: I9c753db619375f4badb85e02826fa3d80673f6b7
Reviewed-on: https://gerrit.chromium.org/gerrit/44962
Reviewed-by: Daniel Erat <derat@chromium.org>
Commit-Queue: Yufeng Shen <miletus@chromium.org>
Tested-by: Yufeng Shen <miletus@chromium.org>