Add disabling of dark resume
To prevent dark resume from breaking suspend autotests, this change adds a pref
that is added for test images. It disables dark resume when this pref is set to
non-zero. This also sets the values of the dark resume sysfs files to disabled
when dark resume is disabled and when DarkResumePolicy is destroyed. This
prevents stale information from restarts and crashes of powerd from persisting.

suspend_stress_test is modified to disable dark resume since it can be run on a
dev image.

dark_resume_policy_unittest is modified to test that this new functionality
works.

BUG=chromium-os:39707
TEST=suspend_stress_test, power_UiResume, power_Resume

Change-Id: I3f31086e5138d317d9293d9130fbec5a81678058
Reviewed-on: https://gerrit.chromium.org/gerrit/44996
Reviewed-by: Derek Basehore <dbasehore@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>