shill: WiFiProvider: Add visible networks by security class
When a service is created due to a visible endpoint, create a service
with the generic security parameter (e.g., WPA -> PSK).  This allows
us to set the security of the service once, and not have to re-assign
this based on the visibility of WPA / RSN endpoints that may come
along later.  This will allow wpa_supplicant configurations that we
create based on this service to connect to either type of endpoint.

BUG=chromium-os:39813
TEST=Unit tests

Change-Id: I1da4224c73f1be3aad49ce57e3939149ec3f2668
Reviewed-on: https://gerrit.chromium.org/gerrit/45121
Tested-by: Ashok Nagarajan <asnagarajan@chromium.org>
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>