shill: Pass down the BSSID of the destination to crypto_util
This will allow the crypto_util shim to verify that the device listed in
the certificate matches the device shill believes to be the current WiFi
service.  There was some debate around what would eventually end up in
the certificate and this was the final result.

TEST=None
BUG=chromium-os:39056

Change-Id: I83e1a8a12918f6eef6e531959bc485c22ac62cba
Reviewed-on: https://gerrit.chromium.org/gerrit/45245
Tested-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Christopher Wiley <wiley@chromium.org>