ipsec_manager: Be more careful when shutting down
The ipsec manager was sloppy about killing the child starter and
charon processes, in the sense that killing one may perturb the
state (notably the pid file) of the other in ways which make it
difficult to find and kill both processes.  To fix this, create
a "Daemon" object which can hold the pid-file and process state
simultaneously so we can garner state about the running processes
in one pass, and kill them in a separate pass.  This also cleans
up some questionable code, like re-using the starter Process
instance for killing both starter and charon processes during
startup.  It also fixes references to base::FilePath in the
various files in this project where it is used, and uses a
ScopedTempDir for temporary files used in unit tests.

BUG=chromium-os:36959
TEST=Unit tests, manual: start and stop an l2tpipsec connection
repeately using connect-vpn / disconnect-service, and ensure that
no PID files or processes are left over for charon or starter.

Change-Id: I59c2cdabb9b99fc9fc54cb25bab029f8573b8c26
Reviewed-on: https://gerrit.chromium.org/gerrit/45334
Commit-Queue: Paul Stewart <pstew@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>