power: Miscellaneous testing-related cleanup.
This updates common/testrunner.cc to initialize the base
library's command-line and logging code so that GTEST_ARGS
can be used to pass logging- and testing-related flags when
emerging the package with FEATURES=test (e.g. it can be set
to a string that includes --gtest_filter and --vmodule).

It also makes minor additions to several stub classes in
anticipation of a larger change:
BacklightStub::clear_resume_level() resets the
previously-set resume level, and DisplayPowerSetterStub
counts the number of times that SetDisplayPower() has been
called.

BUG=none
TEST=builds and unit tests still pass

Change-Id: I673f5f0dce67168b8d62b4d30c1237abd023f89e
Reviewed-on: https://gerrit.chromium.org/gerrit/45342
Commit-Queue: Daniel Erat <derat@chromium.org>
Tested-by: Daniel Erat <derat@chromium.org>
Reviewed-by: Ryan Harrison <rharrison@chromium.org>