Add embedded upstart job to emit login-prompt-visible
The login-prompt-visible upstart signal is normally emitted by the
browser when it's got the sign-in UI up on the screen.  Embedded
devices don't have the browser or, at least so far, the standard
sign-in UI.  Therefore, we need to emit this when the terminal is
ready to go instead.

BUG=chromium:189662
TEST=build_packages --board daisy-drm, check for /build/daisy-drm/etc/init/login-prompt-visible.conf

Change-Id: Iaa4ef079e33848445b0b4d8de7b13509fc07bdd1
Reviewed-on: https://gerrit.chromium.org/gerrit/45347
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Commit-Queue: Chris Masone <cmasone@chromium.org>